ml-service/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pandas==2.1.3
numpy==1.26.2
scikit-learn==1.3.2
scikit-surprise==1.1.3
sqlalchemy==2.0.23
pymysql==1.1.0
python-dotenv==1.0.0